chore: http transport service refactoring
