+++ common-lib/utils/k8s/bean.go
@@ -27,12 +27,7 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	utilnet "k8s.io/apimachinery/pkg/util/net"
 	"k8s.io/client-go/rest"
-	"log"
-	"net"
-	"net/http"
-	"time"
 )
 
 type ClusterConfig struct {
@@ -168,15 +163,6 @@ func GetResourceKey(obj *unstructured.Unstructured) ResourceKey {
 	return NewResourceKey(gvk.Group, gvk.Kind, obj.GetNamespace(), obj.GetName())
 }
 
-type CustomK8sHttpTransportConfig struct {
-	UseCustomTransport  bool `env:"USE_CUSTOM_HTTP_TRANSPORT" envDefault:"false"`
-	TimeOut             int  `env:"K8s_TCP_TIMEOUT" envDefault:"30"`
-	KeepAlive           int  `env:"K8s_TCP_KEEPALIVE" envDefault:"30"`
-	TLSHandshakeTimeout int  `env:"K8s_TLS_HANDSHAKE_TIMEOUT" envDefault:"10"`
-	MaxIdleConnsPerHost int  `env:"K8s_CLIENT_MAX_IDLE_CONNS_PER_HOST" envDefault:"25"`
-	IdleConnTimeout     int  `env:"K8s_TCP_IDLE_CONN_TIMEOUT" envDefault:"300"`
-}
-
 type LocalDevMode bool
 
 type RuntimeConfig struct {
@@ -189,65 +175,14 @@ func GetRuntimeConfig() (*RuntimeConfig, error) {
 	return cfg, err
 }
 
-func NewCustomK8sHttpTransportConfig() *CustomK8sHttpTransportConfig {
-	customK8sHttpTransportConfig := &CustomK8sHttpTransportConfig{}
-	err := env.Parse(customK8sHttpTransportConfig)
-	if err != nil {
-		log.Println("error in parsing custom k8s http configurations from env : ", "err : ", err)
-	}
-	return customK8sHttpTransportConfig
-}
-
-// OverrideConfigWithCustomTransport
-// overrides the given rest config with custom transport if UseCustomTransport is enabled.
-// if the config already has a defined transport, we don't override it.
-func (impl *CustomK8sHttpTransportConfig) OverrideConfigWithCustomTransport(config *rest.Config) (*rest.Config, error) {
-	if !impl.UseCustomTransport || config.Transport != nil {
-		return config, nil
-	}
-
-	dial := (&net.Dialer{
-		Timeout:   time.Duration(impl.TimeOut) * time.Second,
-		KeepAlive: time.Duration(impl.KeepAlive) * time.Second,
-	}).DialContext
-
-	// Get the TLS options for this client config
-	tlsConfig, err := rest.TLSConfigFor(config)
-	if err != nil {
-		return nil, err
-	}
-
-	transport := utilnet.SetTransportDefaults(&http.Transport{
-		Proxy:               config.Proxy,
-		TLSHandshakeTimeout: time.Duration(impl.TLSHandshakeTimeout) * time.Second,
-		TLSClientConfig:     tlsConfig,
-		MaxIdleConns:        impl.MaxIdleConnsPerHost,
-		MaxConnsPerHost:     impl.MaxIdleConnsPerHost,
-		MaxIdleConnsPerHost: impl.MaxIdleConnsPerHost,
-		DialContext:         dial,
-		DisableCompression:  config.DisableCompression,
-		IdleConnTimeout:     time.Duration(impl.IdleConnTimeout) * time.Second,
-	})
-
-	rt, err := rest.HTTPWrappersForConfig(config, transport)
-	if err != nil {
-		return nil, err
-	}
-
-	config.Transport = rt
-	config.Timeout = time.Duration(impl.TimeOut) * time.Second
-
-	// set default tls config and remove auth/exec provides since we use it in a custom transport.
-	// we already set tls config in the transport
-	config.TLSClientConfig = rest.TLSClientConfig{}
-	config.AuthProvider = nil
-	config.ExecProvider = nil
-
-	return config, nil
-}
-
 var NotFoundError = errors.New("not found")
 
 func IsNotFoundError(err error) bool {
 	return errors.Is(err, NotFoundError)
+}
+
+type JsonPatchType struct {
+	Op    string      `json:"op"`
+	Path  string      `json:"path"`
+	Value interface{} `json:"value"`
 }
